+++ include/constants/battle_intro.h
@@ -1,6 +1,11 @@
 #ifndef GUARD_CONSTANTS_BATTLE_INTRO_H
 #define GUARD_CONSTANTS_BATTLE_INTRO_H
 
+// The letterbox animation in the battle intro starts off at a slow speed, 
+// and then picks up the pace at a very specific threshold.
+#define BATTLE_INTRO_LETTERBOX_RECEDE_SPEED_INITIAL 2 // vanilla: 1
+#define BATTLE_INTRO_LETTERBOX_RECEDE_SPEED_BOOSTED 4 // vanilla: 4
+
 // vanilla: 2
 #define BATTLE_INTRO_COMBATANT_SLIDE_IN_SPEED 3